+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-04-12 16:36:18
-// Commit: test(state): add unit tests for state business logic
+// Updated at: 2026-04-12 22:18:57
+// Commit: feat(state): integrate offline caching for state state
 
 class stateComponentState {
-  final String commitRef = '347';
-  final String lastMessage = 'test(state): add unit tests for state business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-12T16:36:18+05:00');
+  final String commitRef = '358';
+  final String lastMessage = 'feat(state): integrate offline caching for state state';
+  final DateTime timestamp = DateTime.parse('2026-04-12T22:18:57+05:00');
 }